"modified where.cc(.cu,.h) and softmax.cu"

+++ include/cuda/cuda_where.h
@@ -4,7 +4,7 @@
 
 namespace infini {
 void where_kernel(const float *inputx, const float *inputy,
-                  const float *condition, float *output, int nDims,
+                  const int *condition, float *output, int nDims,
                   infini::SmallArray inputxShape,
                   infini::SmallArray inputyShape,
                   infini::SmallArray conditionShape,
